requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.6
langchain-huggingface>=0.0.3
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pypdf>=4.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.9
torch>=2.0.0
transformers>=4.40.0
